% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{CalcMitoPct}
\alias{CalcMitoPct}
\title{Calculate the mitochondrial percentage}
\usage{
CalcMitoPct(mysco, prefix)
}
\arguments{
\item{mysco}{the mysco object}

\item{prefix}{the pattern specifying mitochondrial genes}
}
\value{
a mysco object with mito percents in the metadata
}
\description{
Calculate the mitochondrial percentages
}
