% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPCs.R
\name{CalcPCs}
\alias{CalcPCs}
\title{Calculate the principal components of the cells}
\usage{
CalcPCs(mysco)
}
\arguments{
\item{mysco}{The mysco object}
}
\value{
a mysco object with 50 PCs in the pcs slot.
}
\description{
Calculates the first 50 principal components of the cells and stores them.
}
