% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{FilterData}
\alias{FilterData}
\title{Filter single-cell data}
\usage{
FilterData(mysco, sub, min.cells)
}
\arguments{
\item{mysco}{the mysco object}

\item{sub}{The expression of what should be filtred and how.}

\item{min.cells}{Minimum number of cells a gene should be expressed in for it to be kept.}
}
\value{
A filtered mysco object.
}
\description{
make QC plots
}
\examples{
pbmc <- FilterData(pbmc,sub=features > 200 & features < 2500 & perc.mt < 5,min.cells=3)
}
