% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_variable_genes.R
\name{FindHVGs}
\alias{FindHVGs}
\title{Find highly variable genes for further analysis.}
\usage{
FindHVGs(mysco, nhvgs = 2000)
}
\arguments{
\item{mysco}{The mysco object}

\item{nhvgs}{The number of hvgs required.}
}
\value{
A mysco object with variance stats and a list of HVGs.
}
\description{
Calculate and define HVGs
}
