% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscoClass.R
\docType{class}
\name{mysco}
\alias{mysco}
\title{A header line for your function}
\description{
A class of mycso
}
\section{Slots}{

\describe{
\item{\code{meta}}{A data.frame}

\item{\code{meta,genes}}{A data.frame}

\item{\code{counts}}{a matrix of raw counts as class matrix}

\item{\code{data.norm}}{a matrix of normalised and log transformed data}

\item{\code{data.scale}}{a matrix of scaled and centralised data}

\item{\code{hvgs}}{a vector of gene names}

\item{\code{pcs}}{a matrix of the first 50 components}

\item{\code{clusters}}{a vector of cluster assignments}
}}

