% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalise_and_Scale.R
\name{NormaliseData}
\alias{NormaliseData}
\title{Normalise the data according to total read count per cell}
\usage{
NormaliseData(mysco, scale.factor)
}
\arguments{
\item{mysco}{The mysco object}

\item{scale.factor}{The scale factor which the normalised counts are multipled are.}
}
\value{
a mysco object with normalised data in the data.norm slot.
}
\description{
Normalise the data
}
