% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc.data.R
\docType{data}
\name{pbmc.data}
\alias{pbmc.data}
\title{Single cells from Peripheral Blood Mononuclear Cells (PBMC)}
\format{
An matrix of PBMC data, genes in rows, cells in columns
}
\source{
\href{https://satijalab.org/seurat/articles/pbmc3k_tutorial}{Seurat tutorial}
}
\usage{
data(pbmc.data)
}
\description{
2,700 single cells that were sequenced on the Illumina NextSeq 500, freely available from 10X Genomics
}
\examples{
data(pbmc.data)
pbmc <- CreateMySCO(as.matrix(pbmc.data))
}
\keyword{datasets}
