% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalise_and_Scale.R
\name{ScaleData}
\alias{ScaleData}
\title{Scale the data}
\usage{
ScaleData(mysco)
}
\arguments{
\item{mysco}{The mysco object}
}
\value{
a mysco object with normalised data in the data.norm slot.
}
\description{
Centralise and standardise the means and variances.
}
