% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{show the object contents}
\usage{
\S4method{show}{mysco}(object)
}
\arguments{
\item{object}{the cellexalvr object}
}
\value{
nothing
}
\description{
show the contents of a mysco object
}
\details{
Detail the contents of a mysco object
}
