% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterCells.R
\name{ClusterCells}
\alias{ClusterCells}
\title{Calculate a Leiden clustering of the cells}
\usage{
ClusterCells(mysco, nPC = 10, nK, res = 1)
}
\arguments{
\item{mysco}{The mysco object}

\item{nPC}{The number of PCs to use}

\item{nK}{The number of neigbours}

\item{res}{The resolution of the clustering. Default is 1}
}
\value{
a mysco object with clusters in the cluster slot.
}
\description{
Calculates a Leiden clustering of the cells after construction of an SNN
}
